Rename RouteHandler2 to RouteHandler. Delete redundant classes
