--- src/main/java/httpserver/route/method/GetHandler.java
@@ -1,11 +1,0 @@
-package httpserver.route.method;
-
-import httpserver.http.request.Request;
-import httpserver.http.response.Response;
-
-public class GetHandler implements MethodHandler {
-    @Override
-    public Response handle(Request request, Response response) {
-        return response;
-    }
-}
