+++ src/main/java/httpserver/server/ClientHandler.java
@@ -1,7 +1,6 @@
 package httpserver.server;
 
 import httpserver.route.RouteHandler;
-import httpserver.route.RouteHandler2;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -11,11 +10,11 @@ import java.net.Socket;
 
 class ClientHandler implements Runnable {
     private final Socket socket;
-    private final RouteHandler2 routeHandler;
+    private final RouteHandler routeHandler;
     private BufferedReader input;
     private PrintWriter output;
 
-    ClientHandler(Socket socket, RouteHandler2 routeHandler) {
+    ClientHandler(Socket socket, RouteHandler routeHandler) {
         this.socket = socket;
         this.routeHandler = routeHandler;
     }
