+++ src/main/java/httpserver/server/Server.java
@@ -1,7 +1,6 @@
 package httpserver.server;
 
 import httpserver.route.RouteHandler;
-import httpserver.route.RouteHandler2;
 
 import java.io.IOException;
 import java.net.ServerSocket;
@@ -10,9 +9,9 @@ import java.util.concurrent.Executors;
 public class Server {
 
     private final int port;
-    private final RouteHandler2 routeHandler;
+    private final RouteHandler routeHandler;
 
-    public Server(int port, RouteHandler2 routeHandler) {
+    public Server(int port, RouteHandler routeHandler) {
         this.port = port;
         this.routeHandler = routeHandler;
     }
@@ -34,7 +33,7 @@ public class Server {
         }
     }
 
-    private void startServer(ServerSocket serverSocket, RouteHandler2 routeHandler) {
+    private void startServer(ServerSocket serverSocket, RouteHandler routeHandler) {
         var executor = Executors.newCachedThreadPool();
         new ServerRunner(serverSocket, executor, routeHandler).start();
     }
