+++ src/test/java/httpserver/http/RequestResponseHandlerTest.java
@@ -1,6 +1,6 @@
 package httpserver.http;
 
-import httpserver.route.RouteHandler2;
+import httpserver.route.RouteHandler;
 import httpserver.route.Routes;
 import httpserver.server.RequestResponseHandler;
 import org.junit.jupiter.api.Test;
@@ -20,7 +20,7 @@ class RequestResponseHandlerTest {
         var outputStream = new ByteArrayOutputStream();
         var output = new PrintWriter(outputStream, true);
         var routes = new Routes();
-        var routeHandler = new RouteHandler2(routes);
+        var routeHandler = new RouteHandler(routes);
 
         var requestResponseHandler = new RequestResponseHandler(input, output, routeHandler);
         requestResponseHandler.run();
