--- src/test/java/httpserver/route/requestmethod/UnknownMethodHandlerTest.java
@@ -1,20 +1,0 @@
-package httpserver.route.requestmethod;
-
-import httpserver.http.request.RequestFactory;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class UnknownMethodHandlerTest {
-
-    @Test
-    void returnsStatusCode405WhenRequestMethodDoesNotExist() {
-        var rawRequest = "DELETE /get_with_body HTTP/1.1";
-        var request = RequestFactory.build(rawRequest);
-        var unknownMethodHandler = new UnknownMethodHandler();
-
-        var response = unknownMethodHandler.getResponse(request);
-
-        assertEquals("HTTP/1.1 405 METHOD NOT ALLOWED\r\n\r\n", response.toString());
-    }
-}
